+++ builder/src/configure.py
@@ -4,8 +4,9 @@
 from pathlib import Path
 from typing import Any, Dict, Tuple, TypeVar, Union
 
-from brbpkg.constants import WIN_CURSORS_CFG, WIN_DELAY, X_CURSORS_CFG, X_DELAY
 from clickgen.util import PNGProvider
+
+from .constants import WIN_CURSORS_CFG, WIN_DELAY, X_CURSORS_CFG, X_DELAY
 
 X = TypeVar("X")
